refactor: move common types definition to upper layer
